优化线程初始化等待

+++ src/main/java/com/acgist/snail/downloader/torrent/TorrentDownloader.java
@@ -46,7 +46,7 @@ public class TorrentDownloader extends Downloader {
 	public void download() throws IOException {
 		while(ok()) {
 			synchronized (downloadLock) {
-				ThreadUtils.wait(downloadLock, Duration.ofSeconds(Long.MAX_VALUE));
+				ThreadUtils.wait(downloadLock, Duration.ofSeconds(Integer.MAX_VALUE));
 			}
 		}
 	}
